refactor(multiple response): rename folder name
