+++ client/app/bundles/course/assessment/question/multiple-responses/commons/translationAdapter.tsx
